+++ CHANGELOG.md
@@ -5,7 +5,25 @@ All notable changes to this project will be documented in this file.
 <a name="unreleased"></a>
 ## [Unreleased]
 
-- Update docs + lock versions and add changelog
+- Remove lb name's postfix and update examples
+
+
+<a name="1.2.2"></a>
+## [1.2.2] - 2020-11-09
+
+- Update module to remove 0.14 limit ([#4](https://github.com/umotif-public/terraform-aws-alb/issues/4))
+
+
+<a name="1.2.1"></a>
+## [1.2.1] - 2020-08-05
+
+- Improve module versions + add alb access logs example ([#3](https://github.com/umotif-public/terraform-aws-alb/issues/3))
+
+
+<a name="1.2.0"></a>
+## [1.2.0] - 2020-04-21
+
+- Feature/updates ([#2](https://github.com/umotif-public/terraform-aws-alb/issues/2))
 - add git hooks and update docs
 - update docs
 
@@ -34,7 +52,10 @@ All notable changes to this project will be documented in this file.
 - Add default redirect and nlb example
 
 
-[Unreleased]: https://github.com/umotif-public/terraform-aws-alb/compare/1.1.0...HEAD
+[Unreleased]: https://github.com/umotif-public/terraform-aws-alb/compare/1.2.2...HEAD
+[1.2.2]: https://github.com/umotif-public/terraform-aws-alb/compare/1.2.1...1.2.2
+[1.2.1]: https://github.com/umotif-public/terraform-aws-alb/compare/1.2.0...1.2.1
+[1.2.0]: https://github.com/umotif-public/terraform-aws-alb/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/umotif-public/terraform-aws-alb/compare/1.0.4...1.1.0
 [1.0.4]: https://github.com/umotif-public/terraform-aws-alb/compare/1.0.3...1.0.4
 [1.0.3]: https://github.com/umotif-public/terraform-aws-alb/compare/1.0.2...1.0.3